+++ src/test/java/org/apiphany/security/ssl/client/CipherSuites.java
@@ -6,8 +6,10 @@ import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apiphany.io.Int16;
 import org.apiphany.json.JsonBuilder;
 import org.apiphany.lang.ByteSizeable;
+import org.apiphany.security.tls.TLSObject;
 
 public class CipherSuites implements TLSObject {