+++ src/test/java/org/apiphany/security/ssl/client/RecordContentType.java
@@ -2,13 +2,15 @@ package org.apiphany.security.ssl.client;
 
 import java.util.Map;
 
+import org.apiphany.io.Int8;
+import org.apiphany.security.tls.TLSObject;
 import org.morphix.lang.Enums;
 
 public enum RecordContentType implements TLSObject {
 
 	CHANGE_CIPHER_SPEC((byte) 0x14, FromFunction.ignoreSize(ChangeCipherSpec::from)),
 	ALERT((byte) 0x15, FromFunction.ignoreSize(Alert::from)),
-	HANDSHAKE((byte) 0x16, FromFunction.ignoreSize(TLSHandshake::from)),
+	HANDSHAKE((byte) 0x16, FromFunction.ignoreSize(Handshake::from)),
 	APPLICATION_DATA((byte) 0x17, ApplicationData::from),
 	HEARTBEAT((byte) 0x18, null);
 
@@ -35,9 +37,9 @@ public enum RecordContentType implements TLSObject {
 
 	public static RecordContentType from(final TLSObject tlsObject) {
 		return switch(tlsObject) {
-			case TLSHandshakeBody tlsHandshakeObject -> HANDSHAKE;
+			case TLSHandshakeBody tlsHandshakeBody -> HANDSHAKE;
 			case ChangeCipherSpec changeCipherSpec -> CHANGE_CIPHER_SPEC;
-			case Encrypted encryptedFinished -> HANDSHAKE;
+			case Encrypted encrypted -> HANDSHAKE;
 			case ApplicationData applicationData -> APPLICATION_DATA;
 			default -> throw new UnsupportedOperationException("Unknown TLS object type: " + tlsObject.getClass());
 		};