+++ src/test/java/org/apiphany/security/ssl/client/TLSHandshakeBody.java
@@ -1,7 +1,9 @@
 package org.apiphany.security.ssl.client;
 
+import org.apiphany.security.tls.TLSObject;
+
 public interface TLSHandshakeBody extends TLSObject {
 
-	HandshakeType type();
+	HandshakeType getType();
 
 }